+++ mysql_rep_admin.py
@@ -420,7 +420,6 @@ def chk_slv_time(master, slaves, **kwargs):
 
     slaves = list(slaves)
     frmt = kwargs.get("form", "standard")
-    mail = kwargs.get("mail", None)
 
     if frmt == "JSON":
         outdata = {"application": "MySQL Replication",
@@ -445,21 +444,41 @@ def chk_slv_time(master, slaves, **kwargs):
 
     if frmt == "JSON":
         outdata = add_miss_slaves(master, outdata)
-        jdata = json.dumps(outdata, indent=4)
-        mongo_cfg = kwargs.get("class_cfg", None)
-        db_tbl = kwargs.get("db_tbl", None)
-        ofile = kwargs.get("ofile", None)
+        _process_json(outdata, **kwargs)
 
-        if mongo_cfg and db_tbl:
-            db, tbl = db_tbl.split(":")
-            mongo_libs.ins_doc(mongo_cfg, db, tbl, outdata)
 
-        if ofile:
-            gen_libs.write_file(ofile, "w", jdata)
+def _process_json(outdata, **kwargs):
 
-        if mail:
-            mail.add_2_msg(jdata)
-            mail.send_mail()
+    """Function:  _process_json
+
+    Description:  Private function for chk_slv_time().  Process JSON data.
+
+    Arguments:
+        (input) outdata -> JSON document of Check Slave Time output.
+        (input) **kwargs:
+            ofile -> file name - Name of output file.
+            db_tbl -> database:collection - Name of db and collection.
+            class_cfg -> Server class configuration settings.
+            mail -> Mail instance.
+
+    """
+
+    jdata = json.dumps(outdata, indent=4)
+    mongo_cfg = kwargs.get("class_cfg", None)
+    db_tbl = kwargs.get("db_tbl", None)
+    ofile = kwargs.get("ofile", None)
+    mail = kwargs.get("mail", None)
+
+    if mongo_cfg and db_tbl:
+        db, tbl = db_tbl.split(":")
+        mongo_libs.ins_doc(mongo_cfg, db, tbl, outdata)
+
+    if ofile:
+        gen_libs.write_file(ofile, "w", jdata)
+
+    if mail:
+        mail.add_2_msg(jdata)
+        mail.send_mail()
 
 
 def _process_time_lag(slv, time_lag, name, frmt, **kwargs):